// src/app/(crm)/crm/event-categories/types.ts
import { IContractTemplate } from '../contract-templates/type';

export interface ICustomIcon {
  id: string;
  name: string;
  svg_code: string;
  preview_color?: string;
}

export interface IEventCategory {
  id: string;
  name: string;
  color: string;
  description: string | null;
  is_active: boolean;
  icon_id: string | null;
  icon?: ICustomIcon;
  contract_template_id: string | null;
  contract_template?: IContractTemplate;
  created_at: string;
  updated_at: string;
}